Update patch set 5
Patch Set 5:

(6 comments)

Thanks Michal

I think I got the global view.
The heritage and modularity looks great and this readable even for integrator people 😊

the settings is param:value, could it be json to set the composition of VSP as mentioned by Thierry. I assume some service could be constituted of several VSPs, do we suppose that everything is in 1 single zip? or shall we plan a service view of this composition?

at the end we should see if we can leverage the structure and replay the tests we started to work on:
- basic_vm: similar to vfw
- freeradius_nbi includes the creation and destruction of network for instance - but maybe they shall be created through ONAP + use of NBI
- clearwater_vims: sounds more Telco but it is based on heat including very old ubuntu images

I understand that we do no need not have onboard/instantiate/e2e due to the heritage of the structure
we could call only the onboard class if we just want to onboard
for e2e we just need to call instantiate which inherits from onboard (we can only instantiate things we onboard)

let discuss it tomorrow and see how to split the load

Patch-set: 5
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>