Prepare python package to integrate it in xtesting docker
- creation of entrypoint
- include templates in the package
- complete requirements

Open questions for integration python reviewers:
- I was not able with the MANIFEST to use the template directory  until
I put it in the package tree. I am not sure it is very nice..
Maybe it is not the right way, we could simply copy the templates in the
docker rather than including them as part of the python package..

- When I install the package using pip install git+https, it does not take
into account the requirements.txt (I think it is normal for the package, it
shall be part of the setup.py, but how to indicate a develop branch?
Until we release a version for pythonsdk I need to use the develop branch

Note: I retested the ubuntu16 (onboarding/deployment/instantiation) on
the Daily Master

Issue-ID: TEST-258

Signed-off-by: mrichomme <morgan.richomme@orange.com>
Change-Id: I26f3e959b9c5e341ab197a6c519ca87e31921921