Update patch set 2
Patch Set 2:

It looks fine. With sys.path[-1], it is the complete src path of python package that is considered.
The drawback is that all inputs for test to run or cnf/vnf to onboard/instantiate will have to be copied under this python path.
Another solution would be to consider these inputs as variable file outside the package. It looks more flexible

Patch-set: 2
Reviewer: Gerrit User 5622 <5622@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>