Update patch set 3
Patch Set 3: Code-Review+1

for the moment, Thierry needs to use a MR branch of the develop branch. The k8splugin was not integrated in the 7.0.0 because some tests are missing in pythonsdk, especially the functional tests.
The quality requirements for pythonsdk are high and must remain as such to ensure the maintainability of the SDK.

@Michal C: do you have any plan to complete the testing of the feature in pythonsdk, so it would be possible to integrate the code in a 7.0.1 and avoid building a temp docker to include this use case (which can be a temp workaround)

Patch-set: 3
Reviewer: Gerrit User 5553 <5553@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1