Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> maybe a stupid question...
> 
> as the sim is launched as a docker and you mention that most part of the sim remain unchanged, would it be not better to avoid duplicating the code, improve/adapt it directly in the original repo, then build a docker image directly usable by the test from the original repo published on the nexus
> 
> I do not know if we could think of a wrapper at pythonsdk-test level to be able to launch any kind of simulator as a docker aside of the tests.

That's a great question Morgan! Avoiding duplicate code would be good. The reason it was added as simulator_source/ folder is that the pre-built image weighted 107MB and it of course made no sense to upload it to this repo. The simulator_source/ however weighted 1MB and could be built in place quite fast and that's why it's in this commit. After that I was looking for the pre-built image in Nexus and haven't found any with the same name (mass-pnf-sim).

We could either (1) pre-build it, upload it to Nexus to have a stable version there and then only pull the image for the our purposes or alternatively (2) we could pull the relevant repo during execution and build it in place. Is the option #1 what you meant? 

The wrapper of course could do it! We'd then need to think about what the starting point for the wrapper would be (like pulling from Nexus, for example). I think for the wrapper then it'd need to be quite universal to support a variety of possible simulators (for example, pull/run/remove). But then again: different simulators would need custom bootstrap and other custom functionalities (like, trigger event register to VES). How do you see such a wrapper?

Patch-set: 3