Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1: Code-Review+2
> > 
> > would make sense to create an entry point to run a test even if it is basic (enrich CBA) as part of the healthcheck..as CDS is for the moment not very well covered
> 
> Two things needs to be done:
>  - blueprintprocessor has to be exposed,
>  - blueprintprocessor bootstrap needs to be called.
> For the second one there is a change in Python ONAP SDK (https://gitlab.com/Orange-OpenSource/lfn/onap/python-onapsdk/-/merge_requests/207/) so it's not a case, but the what about the first one? Am I understand correct that this code should be provided here, as a step? I mean calling using Python k8s client.

Yes I think it would be the best place even if the best solution would be to have an exposed port..but Andreas is waiting for it for a long time
we could have a step exposeCdsNodePortStep that we could deprecated once it will be fixed.
for the basic-cnf, Thierry made some operations at the CI level as a prerequisite, we could imagine to open this node port as this level (xtesting-onap via the ad-hoc ansible role). DT and Orange have the same CI/CD approach but some companies do not have the same CI mechanism so I think it is better to do it here to avoid dependency on the CI even if this dependency is questionable...for the basic_vm/basic_network we rely on a target cloud deployed by the CI (thee is no deployOpenstackTenantStep) but we could imagine if we need to create additional cloud resources to do it here..just need to find the right balance..

Patch-set: 1
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>