Prepare basic_clamp use case
Update of basic_clamp use case
Please note there is a dependency with the onapsdk [1]
due to teh fact that policy_tca (frankfurt) has to be changed
to policy.tca in guilin.

In this version, we still use the clamp authentication
done by certificates.
The last version of clamp supports basic/auth but onapsdk has to be
updated to offer this option, which will better.
Until it is updated, the use case can work using the old cert
authentificaiton method.

This use case has been succesfully tested on
- daily guilin => dcae k8s plugin 3.4.2
- daily master => use the version 3.4.3

Issue-ID: TEST-277

Signed-off-by: mrichomme <morgan.richomme@orange.com>
Change-Id: I4c5da94dd5669f328d124f7a85886d248a9c5a61