Refactor pythonsdk_test settings
Up to now, we are using the settings mechanism of python_onapsdk
so when we need to customize pythonsdk it is not possible
workarounds are needed
For instance in basic clamp the path is needed for some clamp files.

I am reusing the same code but pythonsdk_tests and python_onapsdk are
no more linked by one single env variable.

Issue-ID: TEST-290

Signed-off-by: mrichomme <morgan.richomme@orange.com>
Change-Id: Iadd4fabf5093ef03e4aa6ed945e930870cb46b9d