Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review+1
> 
> ok for me, but the authentication will have to be fixed soon, because the certificate if going to expire this year i think.

yep, I have this task in my todolist => to be done in python-onapsdk
a refactoring of the settings is also needed https://gerrit.onap.org/r/c/testsuite/pythonsdk-tests/+/116479 ..but some tests are needed here as it may break the CI chains..

as basic_clamp is not in the CI for the moment merging this change has no consequences. It simplifies the manual run.

in case of you missed it => demo basic_clamp done during integration meeting end of december: https://youtu.be/EO14Y8AMRog
the goal is to integrate it in CI in honolulu (with basic/auth and no more the certif)

Patch-set: 2