Update patch set 1
Patch Set 1:

It implements only STEP LEVEL 1 from this concept: 
https://wiki.onap.org/display/DW/Pythonsdk-test+wrapper+for+simulators

The rest will be built and expanded on top of this and submitted as patch sets here. Masspnfsimulator case is a great example on which to implement LEVELs 0,2,3 and test everything. If you have any comments on this, I'm happy to hear them!

Patch-set: 1