Update patch set 3
Patch Set 3:

you applied this patch to guilin branch
in Guilin we fixed the version of the lib to 7.3.0 so it is working fine...
https://logs.onap.org/onap-integration/daily/onap_daily_pod4_guilin/02-08-2021_03-02/

the issue is now on the master branch (which is used for master weely/daily and gating CI/CD chains), which used to consume the latest version of python-onapsdk

So I would apply this patch only on master first
and if OK cherry-pick to guilin to backport the change and use the 7.4.0

Regarding the remark from Sylvain, as the pythonsdk-tests are used on master and gating, we need to find a procedure to be more careful on sdk update... We could use the develop branch of pythonsdk-tests to give a try on the openlab, then when it is validated we merge it to master. I need to formalize something here, I will share by mail.

otherwise for the gating, another option would be to use the latest stable (guilin here)?
but if we add a verification through the develop versions of the docker, it shall be safe enough for the chains.

Patch-set: 3
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>