[TEST] Catch k8s connection exceptions
In two steps: ExposeCDSBlueprintprocessorNodePortStep and PnfSimulatorCnfRegisterStep we uses kubernetes Python library to
prepare environment or check the status of pod instantiation. If connection with k8s api gateway can't be established an
exception is raised and no test report is created.
Add try...except block to catch connection error and finish test gently

Issue-ID: TEST-336
Signed-off-by: Michal Jagiello <michal.jagiello@t-mobile.pl>
Change-Id: I1993bec4c886d77645e1a60e0f3d169abeb4e8bd