Update patch set 2
Patch Set 2:

Storing the CBA as zip file is not ideal. Say we want to test a new resource resolution feature - this would require a change in assertions as well as a change to the CBA, but we would not really see what changed in the CBA if we upload a new zip. 

These CBA's also serve as examples of how to use CDS, I refer to them regularly in discussions on CDS functionality.

Maybe we could have the test clone CBA and data-dictionaries from another repo?

Patch-set: 2
Reviewer: Gerrit User 6218 <6218@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>