Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Storing the CBA as zip file is not ideal. Say we want to test a new resource resolution feature - this would require a change in assertions as well as a change to the CBA, but we would not really see what changed in the CBA if we upload a new zip. 
> 
> These CBA's also serve as examples of how to use CDS, I refer to them regularly in discussions on CDS functionality.
> 
> Maybe we could have the test clone CBA and data-dictionaries from another repo?

Agree, that's not an ideal, but in the current flow of tests execution we just use an image which is created in another repository and there we simply use `pip` to install that library. The change you request requires more changes in the repository I mentioned before.

Patch-set: 2