[TEST-356] Do not create pnf simulator in "onap" namespace
As described in ticket - that resource shouldn't be created in onap namespace because if it's not properly deleted it has impact on the security tests results.
Note: That change doesn't require any additional work in lab preparation - the namespace is going to be created by MSB K8S plugin. Namespace cleanup is only needed

Issue-ID: TEST-356
Signed-off-by: Michal Jagiello <michal.jagiello@t-mobile.pl>
Change-Id: Ie7b288ce81b0fab0f746622d863cd48497d739e9