[TEST] Do not create simulators in "onap" namespace
As described in ticket - that resource shouldn't be created in onap namespace because if it's not properly deleted it has impact on the security tests results.
Note: That change doesn't require any additional work in lab preparation - the namespace is going to be created by MSB K8S plugin.

WIP: Locally works fine, on gating there are strange errors (no value in settings)

Issue-ID: TEST-356
Signed-off-by: Michal Jagiello <michal.jagiello@t-mobile.pl>
Change-Id: I0acf6262e3c8c9fbe9113d96e62960ce36ece04c