[TEST] Do not show "No execution start time..." error log message for skipped substeps
If one of the step fails and there are more substeps in the scenario then they won't be executed.
During the report creation in the code the status of these steps were not checked and the confusing
error message was logged.
That change provide additional check if the step was skipped (not executed).

Issue-ID: TEST-368
Signed-off-by: Michal Jagiello <michal.jagiello@t-mobile.pl>
Change-Id: Id1ea95229b2fd5cf3107d423f356f02dab437669