[TEST] PNF macro nad CDS resource resoulution tests are failing on retry
If any of these two tests fail on the first try retry will always fail as well.
It's because of that test's cleanup phase was executed only if an execution phase passed (all substeps).
Because of that simulators was never deleted and second instantiation try fails every time.
That change split test into two phases: execution and cleanup - if execution fails cleanup will be executed despite of that.

Issue-ID: TEST-354
Signed-off-by: Michal Jagiello <michal.jagiello@t-mobile.pl>
Change-Id: Icf4abb1d82c64d1168e6200e94f9e05b54223d9c